# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_import_json_api_v1_topologies_import_json_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_json_api_v1_topologies_import_json_post",
        "type": "object"
      },
      "ContextOut": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeOut"
          },
          "topology": {
            "$ref": "#/components/schemas/TopologyRecord"
          }
        },
        "required": [
          "topology",
          "plan",
          "runtime",
          "events"
        ],
        "title": "ContextOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanOut": {
        "properties": {
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          }
        },
        "required": [
          "plan",
          "diagnostics"
        ],
        "title": "PlanOut",
        "type": "object"
      },
...
      "RuntimeOut": {
        "properties": {
          "active_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeState"
            },
            "title": "Nodes",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "topology_id": {
            "title": "Topology Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "topology_id",
          "status",
          "version",
          "nodes"
        ],
        "title": "RuntimeOut",
        "type": "object"
      },
...
      "TopologyCreate": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "data"
        ],
        "title": "TopologyCreate",
        "type": "object"
      },
      "TopologyRecord": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "engine_state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine State"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "version",
          "created_at",
          "updated_at",
          "data"
        ],
        "title": "TopologyRecord",
        "type": "object"
      },
      "TopologySummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "TopologySummary",
        "type": "object"
      },
      "TopologyUpdate": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "TopologyUpdate",
        "type": "object"
      },
      "ValidateBody": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "ValidateBody",
        "type": "object"
      },
...
      "ValidationResult": {
        "properties": {
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/DiagnosticsSummary"
          }
        },
        "required": [
          "diagnostics",
          "summary"
        ],
        "title": "ValidationResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AE3GIS API",
    "version": "3.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/topologies": {
      "get": {
        "operationId": "list_topologies_api_v1_topologies_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopologySummary"
                  },
                  "title": "Response List Topologies Api V1 Topologies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Topologies",
        "tags": [
          "topologies"
        ]
      },
      "post": {
        "operationId": "create_topology_api_v1_topologies_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopologyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Topology",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/import-json": {
      "post": {
        "operationId": "import_json_api_v1_topologies_import_json_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_json_api_v1_topologies_import_json_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Json",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/validate": {
      "post": {
        "description": "Validate a topology document without saving it.",
        "operationId": "validate_data_api_v1_topologies_validate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Data",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/{topology_id}": {
      "delete": {
        "operationId": "delete_topology_api_v1_topologies__topology_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Topology",
        "tags": [
          "topologies"
        ]
      },
      "get": {
        "operationId": "get_topology_api_v1_topologies__topology_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Topology",
        "tags": [
          "topologies"
        ]
      },
      "put": {
        "operationId": "update_topology_api_v1_topologies__topology_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopologyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Topology",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/{topology_id}/context": {
      "get": {
        "description": "Everything about a topology in one response: record + diagnostics, plan, runtime, recent events.",
        "operationId": "get_context_api_v1_topologies__topology_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context",
        "tags": [
          "topologies"
        ]
      }
    },
...
    "/api/v1/topologies/{topology_id}/events": {
      "get": {
        "operationId": "get_events_api_v1_topologies__topology_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Get Events Api V1 Topologies  Topology Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Events",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/{topology_id}/export": {
      "get": {
        "description": "Portable representations of the lab: JSON spec, Kathara folder, or ContainerLab topology (zips).",
        "operationId": "export_topology_api_v1_topologies__topology_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "labspec",
              "pattern": "^(labspec|kathara|containerlab)$",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Topology",
        "tags": [
          "topologies"
        ]
      }
    },
...
    "/api/v1/topologies/{topology_id}/plan": {
      "get": {
        "description": "The engine-agnostic lab plan: what deploy would instantiate.",
        "operationId": "get_plan_api_v1_topologies__topology_id__plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/{topology_id}/runtime": {
      "get": {
        "description": "Deployment status, the active job (with steps), and per-node state in one call.",
        "operationId": "get_runtime_api_v1_topologies__topology_id__runtime_get",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runtime",
        "tags": [
          "topologies"
        ]
      }
    },
    "/api/v1/topologies/{topology_id}/validate": {
      "post": {
        "operationId": "validate_topology_api_v1_topologies__topology_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "topology_id",
            "required": true,
            "schema": {
              "title": "Topology Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Topology",
        "tags": [
          "topologies"
        ]
      }
    }
  }
}